requirements.txt
boto3~=1.34.116
click~=8.1.7
PyYAML~=6.0.1
botocore~=1.34.116